do not apt update and install if npm exists

+++ frmcompanion/init.sh
@@ -4,8 +4,7 @@ test ! -e Companion && /usr/bin/git clone --depth 1 -b multiarch https://github.
 cd Companion;
 mkdir -p ./bin/map
 test ! -e ./bin/companion && /usr/local/go/bin/go build -o bin/companion main.go
-apt update
-apt install -y nodejs npm
+test ! -e /usr/bin/npm && apt update && apt install -y nodejs npm
 cd ../map
 npm install
 npm run compile
